GBUBasketball/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.17.1
gunicorn==25.3.0
Jinja2==3.1.6
jsonfield==3.2.0
MarkupSafe==3.0.3
packaging==26.1
pillow==12.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
setuptools==82.0.1
sqlparse==0.5.5
swapper==1.4.0
tzdata==2026.1
